Moved FragmentCollection declaration to a separate file.
