+++ src/components/contract/ContractByteCodeSection.vue
@@ -164,7 +164,8 @@ import InfoTooltip from "@/components/InfoTooltip.vue";
 import ContractVerificationDialog from "@/dialogs/verification/ContractVerificationDialog.vue";
 import {AppStorage} from "@/AppStorage";
 import ContractSourceValue from "@/components/values/ContractSourceValue.vue";
-import ContractAbiValue, {FragmentCollection} from "@/dialogs/abi/ContractAbiValue.vue";
+import ContractAbiValue from "@/dialogs/abi/ContractAbiValue.vue";
+import {FragmentCollection} from "@/dialogs/abi/FragmentCollection.ts";
 import {SourcifyResponseItem} from "@/utils/cache/SourcifyCache";
 import DownloadButton from "@/components/DownloadButton.vue";
 import JSZip from "jszip";
