+++ src/dialogs/abi/FragmentCollection.ts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export enum FragmentCollection {
+    ALL = "all",
+    READONLY = "read-only",
+    READWRITE = "read-write",
+    EVENTS = "events",
+    ERRORS = "errors",
+    OTHER = "other",
+}
